pyproject.toml
[tool.poetry]
name = "osmaxx"
version = "1.0.0"
description = ""
authors = ["Geometa Lab <geometalab@ost.ch>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
"geometalab.osm-pbf-file-size-estimation-service" = "^3.0.0"
drf-extensions = "^0.7.1"
djangorestframework-gis = "^0.17"
djangorestframework = "^3.12.4"
django-crispy-forms = "^1.12.0"
django-debug-toolbar = "^3.2.2"
django-downloadview = "^2.1.1"
django-environ = "^0.4.5"
django-extensions = "^3.1.3"
django-filter = "^2.4.0"
django-model-utils = "^4.1.1"
django-secure = "^1.0.2"
djangorestframework-jwt = "^1.11.0"
furl = "^2.1.2"
GeoAlchemy2 = "^0.9.3"
Markdown = "^3.3.4"
memoize = "^1.0.0"
osmium = "^3.2.0"
psycopg2-binary = "^2.9.1"
requests = "^2.26.0"
SQLAlchemy = ">=1.4,<1.4.23"
SQLAlchemy-Utils = "^0.37.8"
sqlalchemy-views = "^0.3.0"
Werkzeug = "^2.0.1"
whitenoise = "^5.3.0"
python-social-auth = {extras = ["django"], version = "^0.3.6"}
sentry-sdk = "^1.3.1"
six = "^1.16.0"
scipy = "^1.7.1"
numpy = "^1.21.2"
ipython = "^7.26.0"
gunicorn = "^20.1.0"
posthog = "^1.4.2"
celery = {extras = ["redis", "auth", "msgpack"], version = "^5.1.2"}
django-celery-results = "^2.2.0"
django-celery-beat = "^2.2.1"
pymemcache = "^3.5.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.4.0"
pytest-cov = "^2.12.1"
pytest-mock = "^3.6.1"
pytest-selenium = "^2.0.1"
flake8 = "^3.9.2"
coverage = "^5.5"
mkdocs = "^1.2.2"
PyHamcrest = "^2.0.2"
vcrpy = "^4.1.1"
requests-mock = "^1.9.3"
Jinja2 = "^3.0.1"
"ruamel.yaml" = "^0.17.10"
pypandoc = "^1.6.3"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
